# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  services.ActivityLog:
    properties:
      action:
        type: string
      admin_name:
        type: string
      created_at:
        type: string
      details:
        type: string
      id:
        type: string
      type:
        type: string
    type: object
  services.ChartDataset:
    properties:
      color:
        type: string
      data:
        items:
          type: number
        type: array
      label:
        type: string
    type: object
  services.DashboardOverview:
    properties:
      active_members:
        type: integer
      commission_rate:
        type: number
      growth_rate:
        type: number
      new_members_today:
        type: integer
      pending_income:
        type: number
      total_income:
        type: number
      total_members:
        type: integer
      total_referrals:
        type: integer
    type: object
...
  services.IncomeChartData:
    properties:
      datasets:
        items:
          $ref: '#/definitions/services.ChartDataset'
        type: array
      labels:
        items:
          type: string
        type: array
    type: object
  services.IncomeDashboardStats:
    properties:
      avg_per_member:
        type: number
      highest_amount:
        type: number
      highest_earner:
        type: string
      pending_payout:
        type: number
      total_distributed:
        type: number
      transactions:
        type: integer
    type: object
...
  services.LevelDistribution:
    properties:
      counts:
        items:
          type: integer
        type: array
      levels:
        items:
          type: integer
        type: array
    type: object
...
  services.MemberDashboardStats:
    properties:
      active:
        type: integer
      by_referral_count:
        additionalProperties:
          type: integer
        type: object
      by_status:
        additionalProperties:
          type: integer
        type: object
      inactive:
        type: integer
      pending:
        type: integer
      suspended:
        type: integer
      total:
        type: integer
    type: object
  services.MemberGrowthChart:
    properties:
      growth:
        items:
          type: number
        type: array
      labels:
        items:
          type: string
        type: array
      members:
        items:
          type: integer
        type: array
    type: object
...
  services.SystemAlert:
    properties:
      actions:
        type: string
      id:
        type: string
      message:
        type: string
      severity:
        type: string
      type:
        type: string
    type: object
...
  services.TopEarner:
    properties:
      direct_count:
        type: integer
      member_id:
        type: string
      member_name:
        type: string
      total_income:
        type: number
    type: object
...
  utils.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Admin API for Rudra management system

    This API provides endpoints for managing Rudra operations including:
    - Admin authentication and authorization
    - Member management
    - Referral tree management
    - Income calculations and distributions
    - Reports and analytics
    - Audit logging
  title: Rudra Admin API
  version: 1.0.0
paths:
...
  /api/v1/dashboard/activity:
    get:
      description: Get recent system activity
      parameters:
      - default: 20
        description: Number of activities
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/services.ActivityLog'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Recent activity
      tags:
      - dashboard
  /api/v1/dashboard/alerts:
    get:
      description: Get system alerts and notifications
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/services.SystemAlert'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: System alerts
      tags:
      - dashboard
  /api/v1/dashboard/charts/growth:
    get:
      description: Get member growth chart data
      parameters:
      - default: yearly
        description: Period (yearly, monthly)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.MemberGrowthChart'
              type: object
      security:
      - Bearer: []
      summary: Member growth chart
      tags:
      - dashboard
  /api/v1/dashboard/charts/income:
    get:
      description: Get income chart data for dashboard
      parameters:
      - default: yearly
        description: Period (yearly, monthly)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.IncomeChartData'
              type: object
      security:
      - Bearer: []
      summary: Income chart data
      tags:
      - dashboard
...
  /api/v1/dashboard/income:
    get:
      description: Get income statistics for the dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.IncomeDashboardStats'
              type: object
      security:
      - Bearer: []
      summary: Income dashboard stats
      tags:
      - dashboard
  /api/v1/dashboard/levels:
    get:
      description: Get member distribution across Rudra levels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.LevelDistribution'
              type: object
      security:
      - Bearer: []
      summary: Level distribution
      tags:
      - dashboard
  /api/v1/dashboard/members:
    get:
      description: Get detailed member statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.MemberDashboardStats'
              type: object
      security:
      - Bearer: []
      summary: Member dashboard stats
      tags:
      - dashboard
  /api/v1/dashboard/overview:
    get:
      description: Get high-level dashboard overview stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.DashboardOverview'
              type: object
      security:
      - Bearer: []
      summary: Dashboard overview
      tags:
      - dashboard
  /api/v1/dashboard/top-earners:
    get:
      description: Get top earning members
      parameters:
      - default: 10
        description: Number of top earners
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/services.TopEarner'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Top earners
      tags:
      - dashboard
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"